+++ vmime/security/digest/messageDigest.hpp
@@ -59,7 +59,7 @@ public:
 	  * @param buffer array of bytes
 	  * @param len number of bytes to use in the buffer
 	  */
-	virtual void update(const byte_t* buffer, const unsigned long len) = 0;
+	virtual void update(const byte_t* buffer, const size_t len) = 0;
 
 	/** Updates the digest using the specified array of bytes,
 	  * starting at the specified offset.
@@ -69,8 +69,8 @@ public:
 	  * @param len number of bytes to use, starting at offset
 	  */
 	virtual void update(const byte_t* buffer,
-	                    const unsigned long offset,
-	                    const unsigned long len) = 0;
+	                    const size_t offset,
+	                    const size_t len) = 0;
 
 	/** Completes the hash computation by performing final operations
 	  * such as padding.
@@ -88,22 +88,22 @@ public:
 	  * then finalize().
 	  */
 	virtual void finalize(const byte_t* buffer,
-	                      const unsigned long len) = 0;
+	                      const size_t len) = 0;
 
 	/** Completes the hash computation by performing final operations
 	  * such as padding. This is equivalent to calling update() and
 	  * then finalize().
 	  */
 	virtual void finalize(const byte_t* buffer,
-	                      const unsigned long offset,
-	                      const unsigned long len) = 0;
+	                      const size_t offset,
+	                      const size_t len) = 0;
 
 	/** Returns the length of the hash.
 	  * This is the length of the array returned by getDigest().
 	  *
 	  * @return length of computed hash
 	  */
-	virtual int getDigestLength() const = 0;
+	virtual size_t getDigestLength() const = 0;
 
 	/** Returns the hash, as computed by the algorithm.
 	  * You must call finalize() before using this function, or the
